+++ packages/authhero/test/helpers/test-server.ts
@@ -1,0 +1,36 @@
+import { Kysely, SqliteDialect } from "kysely";
+import SQLite from "better-sqlite3";
+import { DataAdapters } from "@authhero/adapter-interfaces";
+import createAdapters, {
+  Database,
+  migrateToLatest,
+} from "@authhero/kysely-adapter";
+import { init } from "../../src";
+
+type getEnvParams = {
+  testTenantLanguage?: string;
+  emailValidation?: "enabled" | "enforced" | "disabled";
+};
+
+export async function getTestServer(args: getEnvParams = {}) {
+  const dialect = new SqliteDialect({
+    database: new SQLite(":memory:"),
+  });
+
+  // Don't use getDb here as it will reuse the connection
+  const db = new Kysely<Database>({ dialect: dialect });
+
+  // await migrateToLatest(dialect, false, db);
+
+  const data: DataAdapters = createAdapters(db);
+
+  const env = {
+    data,
+  };
+
+  const apps = init({ dataAdapter: data });
+  return {
+    ...apps,
+    env,
+  };
+}