+++ packages/authhero/test/routes/oauth2/well-known.spec.ts
@@ -1,0 +1,28 @@
+import { describe, it, expect } from "vitest";
+import { testClient } from "hono/testing";
+import { getTestServer } from "../../helpers/test-server";
+import { jwksKeySchema } from "@authhero/adapter-interfaces";
+
+describe("jwks", () => {
+  it("should return a list with the test certificate", async () => {
+    const { oauthApp, env } = await getTestServer();
+    const client = testClient(oauthApp, env);
+
+    const response = await client[".well-known"]["jwks.json"].$get(
+      {
+        param: {},
+      },
+      {
+        headers: {
+          "tenant-id": "tenantId",
+        },
+      },
+    );
+
+    expect(response.status).toBe(200);
+
+    const body = await response.json();
+    const jwks = jwksKeySchema.parse(body);
+    expect(jwks.keys.length).toBe(0);
+  });
+});