+++ packages/drizzle/src/index.ts
@@ -1,7 +1,3 @@
-import { tmp } from "@authhero/adapter-interfaces";
-
 export default function create() {
-  console.log(tmp);
-
   throw new Error("Not implemented");
 }