+++ packages/kysely/src/migrate/migrations/2024-10-19T07:47:00_prompt.ts
@@ -1,0 +1,16 @@
+import { Kysely } from "kysely";
+import { Database } from "@authhero/kysely-adapter";
+
+export async function up(db: Kysely<Database>): Promise<void> {
+  await db.schema
+    .alterTable("logins")
+    .addColumn("authParams_prompt", "varchar(16)")
+    .execute();
+}
+
+export async function down(db: Kysely<Database>): Promise<void> {
+  await db.schema
+    .alterTable("logins")
+    .dropColumn("authParams_prompt")
+    .execute();
+}