+++ packages/create-authhero/CHANGELOG.md
@@ -1,5 +1,11 @@
 # create-authhero
 
+## 0.1.2
+
+### Patch Changes
+
+- Expose the migration script for kysely and add authhero test
+
 ## 0.1.1
 
 ### Patch Changes